@@ -2,7 +2,7 @@ package com.vegvisir.application;
 
 import android.app.Application;
 import android.content.Context;
-import android.support.v4.util.Pair;
+import androidx.core.util.Pair;
 
 import com.google.protobuf.ByteString;
 import com.isaacsheff.charlotte.proto.Block;
@@ -73,8 +73,6 @@ public class VegvisirInstanceV1 implements VegvisirInstance, NewBlockListener, R
      * The singleton instance.
      */
     private static VegvisirInstanceV1 instance;
-
-
 
     public static synchronized VegvisirInstance getInstance(Context applicationContext) {
         if (instance == null) {
